requirements.txt
datasets[audio]
sounddevice
numpy
pandas
pydub
matplotlib
seaborn
pyinstrument
jiwer
tensorflow
keras
mltu
tf2onnx
playsound
transformers
pysubs2
tqdm
pyspellchecker

--index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio